Optimize imports and reformat code.

+++ src/me/TomTheDeveloper/stats/MySQLDatabase.java
@@ -2,8 +2,10 @@ package me.TomTheDeveloper.stats;
 
 import org.bukkit.plugin.java.JavaPlugin;
 
-import java.lang.String;import java.sql.Connection;import java.sql.ResultSet;import java.sql.SQLException;import java.sql.Statement;
-import java.util.HashMap;
+import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.UUID;
@@ -12,7 +14,8 @@ public class MySQLDatabase {
 
     public MySQLConnectionManager manager;
     JavaPlugin plugin;
-    public MySQLDatabase(JavaPlugin javaPlugin){
+
+    public MySQLDatabase(JavaPlugin javaPlugin) {
         this.plugin = javaPlugin;
         this.manager = new MySQLConnectionManager(plugin);
         plugin.getLogger().info("Configuring connection pool...");
@@ -21,7 +24,7 @@ public class MySQLDatabase {
 
         try {
             Connection connection = manager.getConnection();
-            if(connection == null){
+            if (connection == null) {
                 System.out.print("CONNECTION TO DATABASE FAILED!");
                 return;
             }
@@ -45,59 +48,59 @@ public class MySQLDatabase {
     }
 
 
-    public void executeUpdate(String query){
+    public void executeUpdate(String query) {
         try {
             Connection connection = manager.getConnection();
             Statement statement = connection.createStatement();
             statement.executeUpdate(query);
             manager.closeConnection(connection);
         } catch (SQLException e) {
-            plugin.getLogger().warning("Failed to execute update: "+query);
+            plugin.getLogger().warning("Failed to execute update: " + query);
         }
 
     }
 
-    public ResultSet executeQuery(String query){
+    public ResultSet executeQuery(String query) {
         try {
             Connection connection = manager.getConnection();
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             manager.closeConnection(connection);
             return rs;
-        } catch (SQLException exception){
+        } catch (SQLException exception) {
             exception.printStackTrace();
-            plugin.getLogger().warning("Failed to execute request: "+query);
+            plugin.getLogger().warning("Failed to execute request: " + query);
             return null;
         }
     }
 
-    public void insertPlayer(String UUID){
-        executeUpdate("INSERT INTO playerstats (UUID,xp) VALUES ('"+UUID+"',0)");
+    public void insertPlayer(String UUID) {
+        executeUpdate("INSERT INTO playerstats (UUID,xp) VALUES ('" + UUID + "',0)");
     }
 
-    public void closeDatabase(){
+    public void closeDatabase() {
         manager.shutdownConnPool();
     }
 
 
-    public void addStat(String UUID, String stat){
-        addStat(UUID, stat,1);
+    public void addStat(String UUID, String stat) {
+        addStat(UUID, stat, 1);
     }
 
-    public void addStat(String UUID, String stat, int amount){
-        executeUpdate("UPDATE playerstats SET "+stat+"="+stat+"+"+amount+" WHERE UUID='"+UUID+"'");
+    public void addStat(String UUID, String stat, int amount) {
+        executeUpdate("UPDATE playerstats SET " + stat + "=" + stat + "+" + amount + " WHERE UUID='" + UUID + "'");
     }
 
-    public void setStat(String UUID, String stat, int number){
-        executeUpdate("UPDATE playerstats SET "+stat+"="+number+" WHERE UUID='"+UUID+"';");
+    public void setStat(String UUID, String stat, int number) {
+        executeUpdate("UPDATE playerstats SET " + stat + "=" + number + " WHERE UUID='" + UUID + "';");
     }
 
-    public int getStat(String UUID, String stat){
-        ResultSet set = executeQuery("SELECT "+stat+" FROM playerstats WHERE UUID='"+UUID+"'");
+    public int getStat(String UUID, String stat) {
+        ResultSet set = executeQuery("SELECT " + stat + " FROM playerstats WHERE UUID='" + UUID + "'");
         try {
-            if(!set.next())
+            if (!set.next())
                 return 0;
-            return(set.getInt(1));
+            return (set.getInt(1));
         } catch (SQLException e) {
             e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
             return 0;
@@ -105,20 +108,18 @@ public class MySQLDatabase {
 
     }
 
-    public Map<UUID,Integer> getColumn(String stat){
-        ResultSet set = executeQuery("SELECT UUID, "+stat+" FROM playerstats ORDER BY "+stat+" DESC;");
+    public Map<UUID, Integer> getColumn(String stat) {
+        ResultSet set = executeQuery("SELECT UUID, " + stat + " FROM playerstats ORDER BY " + stat + " DESC;");
         Map<java.util.UUID, java.lang.Integer> column = new LinkedHashMap<UUID, Integer>();
         try {
-            while(set.next()){
-                column.put(java.util.UUID.fromString(set.getString("UUID")), (java.lang.Integer) set.getInt(stat));
+            while (set.next()) {
+                column.put(java.util.UUID.fromString(set.getString("UUID")), set.getInt(stat));
             }
         } catch (SQLException e) {
             e.printStackTrace();
         }
         return column;
     }
-
-
 
 
 }
